Move flight2 -> flight
