+++ fixtures/flight/.gitignore
@@ -11,7 +11,6 @@
 # production
 /build
 /dist
-.eslintcache
 
 # misc
 .DS_Store
