+++ fixtures/flight/public/favicon.ico
