+++ fixtures/flight/public/logo192.png
