+++ fixtures/flight/public/logo512.png
