+++ fixtures/flight/public/manifest.json
