+++ fixtures/flight/scripts/test.js
@@ -21,7 +21,7 @@ let argv = process.argv.slice(2);
 
 function isInGitRepository() {
   try {
-    execSync('git rev-parse --is-inside-work-tree', {stdio: 'ignore'});
+    execSync('git rev-parse --is-inside-work-tree', { stdio: 'ignore' });
     return true;
   } catch (e) {
     return false;
@@ -30,7 +30,7 @@ function isInGitRepository() {
 
 function isInMercurialRepository() {
   try {
-    execSync('hg --cwd . root', {stdio: 'ignore'});
+    execSync('hg --cwd . root', { stdio: 'ignore' });
     return true;
   } catch (e) {
     return false;
@@ -47,5 +47,6 @@ if (
   const hasSourceControl = isInGitRepository() || isInMercurialRepository();
   argv.push(hasSourceControl ? '--watch' : '--watchAll');
 }
+
 
 jest.run(argv);
