--- fixtures/flight2/loader/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
