--- fixtures/flight2/src/Container.js
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-export default function Container({children}) {
-  return <div>{children}</div>;
-}
