--- fixtures/flight2/src/Counter.client.js
@@ -1,12 +1,0 @@
-import * as React from 'react';
-
-import Container from './Container.js';
-
-export function Counter() {
-  const [count, setCount] = React.useState(0);
-  return (
-    <Container>
-      <button onClick={() => setCount(c => c + 1)}>Count: {count}</button>
-    </Container>
-  );
-}
