--- fixtures/flight2/src/Counter2.client.js
@@ -1,1 +1,0 @@
-export * from './Counter.client.js';
